+++ app/models/users.m.js
@@ -1,4 +1,4 @@
-const connection = require("../../connection");
+const connection = require("../../db/connection");
 
 exports.selectUsers = () => {
   return connection
